requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.3
pyyaml>=6.0.1
tqdm>=4.6.0
yt-dlp>=2025.02.19
lxml>=5.1.0 

# Selenium and ChromeDriver management
selenium>=4.18.1
webdriver-manager>=4.0.1
cloudscraper

# SMB support
smbprotocol>=1.13.0
pysmb

# WebDAV support
webdavclient3>=3.14.6

# Logging and utilities
loguru>=0.7.2
termcolor>=2.5.0
rich>=13.7
art>=6.4

# API server (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
